requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
glcontext==2.5.0
h11==0.16.0
idna==3.10
moderngl==5.10.0
multipledispatch==1.0.0
numpy==2.3.4
pillow==12.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pyrr==0.10.3
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
